Factor out requires private user url matcher test
