+++ tests/Functional/Services/UrlMatcher/UrlMatcherTest.php
@@ -8,13 +8,6 @@ use App\Tests\Functional\AbstractBaseTestCase;
 
 class UrlMatcherTest extends AbstractBaseTestCase
 {
-    use ValidUserUrlMatcherMatchNotExpectedDataProviderTrait;
-
-    /**
-     * @var UrlMatcher
-     */
-    private $requiresPrivateUserUrlMatcher;
-
     /**
      * @var UrlMatcher
      */
@@ -24,41 +17,9 @@ class UrlMatcherTest extends AbstractBaseTestCase
     {
         parent::setUp();
 
-        $this->requiresPrivateUserUrlMatcher = self::$container->get(
-            'simplytestable.web_client.requires_private_user_url_matcher'
-        );
-
         $this->requiresValidTestOwnerUrlMatcher = self::$container->get(
             'simplytestable.web_client.requires_valid_test_owner_url_matcher'
         );
-    }
-
-    /**
-     * @dataProvider requiresPrivateUserUrlMatcherMatchExpectedDataProvider
-     */
-    public function testRequiresPrivateUserUrlMatcherMatchExpected(string $path)
-    {
-        $this->assertTrue($this->requiresPrivateUserUrlMatcher->match($path));
-    }
-
-    public function requiresPrivateUserUrlMatcherMatchExpectedDataProvider(): array
-    {
-        return [
-            '/account/* [1]' => [
-                'path' => '/account/news-subscriptions/update/',
-            ],
-            '/account/* [2]' => [
-                'path' => '/account/',
-            ],
-        ];
-    }
-
-    /**
-     * @dataProvider requiresValidUserUrlMatcherMatchNotExpectedDataProvider
-     */
-    public function testRequiresPrivateUserUrlMatcherMatchNotExpected(string $path)
-    {
-        $this->assertFalse($this->requiresPrivateUserUrlMatcher->match($path));
     }
 
     /**
